# [wip] tests: benchmark degradation of write throughput as a function of splits