Touches https://github.com/cockroachdb/cockroach/issues/136366

It's not degrading very aggressively in this benchmark:

```
$ benchstat -col /splits bench.txt
goos: linux
goarch: amd64
pkg: github.com/cockroachdb/cockroach/pkg/sql/tests
cpu: Intel(R) Xeon(R) CPU @ 2.80GHz
                           │      0      │              10               │              100              │                1000                │
                           │   sec/op    │   sec/op     vs base          │   sec/op     vs base          │   sec/op     vs base               │
KV/Insert/Native/rows=1-24   31.01µ ± 1%   30.90µ ± 1%  ~ (p=0.358 n=20)   31.32µ ± 1%  ~ (p=0.355 n=20)   31.80µ ± 1%  +2.55% (p=0.000 n=20)
```

However, since this test is single-node, there's less raft scheduling overhead (since fewer round-trips through raft are necessary to commit something).

Epic: CRDB-42584

Release note: None